line-range: fix infinite loop bug with degenerate '$' regex
When the -L argument to "git log" is passed a regular expression
that matches an empty string at the end of a line (e.g. "$" as in
"-L :$:line-range.c"), find_funcname_matching_regexp() falls into
an infinite loop.

Modify find_funcname_matching_regexp() to correctly match the entire
line, instead of the zero-width match at eol, and update the loop
condition to prevent an infinite loop in the event of other
undiscovered corner cases.

Originally reported in <https://stackoverflow.com/q/74690545/147356>.

Signed-off-by: Lars Kellogg-Stedman <lars@oddbit.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>